move controller and mailer outside of spree namespace; use haml in template
