+++ app/views/payment_mailer/authorize_payment.html.haml
