+++ app/views/payment_mailer/authorize_payment.text.haml
@@ -1,0 +1,3 @@
+= t('.instructions', distributor: @payment.order.distributor.name, amount: @payment.display_amount)
+
+= main_app.authorize_payment_url(@payment)
