--- app/views/spree/payment_mailer/authorize_payment.text.erb
@@ -1,2 +1,0 @@
-<%= t('.instructions', distributor: @payment.order.distributor.name, amount: @payment.display_amount) %>
-<%= main_app.authorize_payment_url(@payment) %>
